♻️ refactor : 함수 목적에 맞게 파일명 분리 #9
