+++ src/services/user.dto.ts
@@ -8,11 +8,6 @@ interface UserDataProps {
   email: string;
 }
 
-interface SingleDataProps<T> {
-  table: (typeof TABLES)[keyof typeof TABLES];
-  field: string;
-  value: string;
-}
 
 export const postUserData = async (userData: UserDataProps) => {
   const supabase = await createClient();
@@ -34,17 +29,4 @@ export const postUserData = async (userData: UserDataProps) => {
   }
 };
 
-export const getSingleData = async <T>({
-  table,
-  field,
-  value,
-}: SingleDataProps<T>) => {
-  const supabase = await createClient();
-  const { data, error } = await supabase
-    .from(table)
-    .select('*')
-    .eq(field, value)
-    .single<T>();
 
-  return { data, error };
-};
