+++ src/utils/duplicate-validation.ts
@@ -2,8 +2,8 @@
 
 import { TABLES } from '@/constants/tables';
 import { ERROR_MESSAGES } from '@/constants/messages';
-import { getSingleData } from '@/services/user.dto';
 import { User } from '@/types/supabase.type';
+import { getSingleData } from '@/services/common.dto';
 
 export const duplicateEmailValidation = async (email: string) => {
   const { data, error } = await getSingleData<User>({
